requirements.txt
fastapi
uvicorn
SQLAlchemy
alembic
python-dotenv
pydantic
passlib[bcrypt]
python-jose
pydantic-settings
celery 
redis
pydantic[email]